# kbmk_backend-main/storage/api-docs/api-docs.yaml
openapi: 3.0.0
info:
  title: 'KBMK UNMUL API Documentation'
  description: 'Dokumentasi API untuk team Frontend'
  contact:
    email: admin@admin.com
  license:
    name: 'Apache 2.0'
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
  version: 1.0.0
servers:
  -
    url: /api
paths:
...
  /documentations:
    get:
      tags:
        - Documentations
      summary: 'Menampilkan daftar dokumentasi (dengan paginasi)'
      operationId: 5fc3487a79802a1975f5658b46bd9691
      responses:
        200:
          description: Berhasil
          content:
            application/json:
              schema:
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Documentation'
                  links:
                    type: object
                  meta:
                    type: object
                type: object
    post:
      tags:
        - Documentations
      summary: 'Menambah dokumentasi baru'
      description: "Memerlukan autentikasi dan role 'admin'. Mengupload foto."
      operationId: b518a17518d679c1381e7b7c593b3610
      requestBody:
        content:
          multipart/form-data:
            schema:
              required:
                - nama
                - deskripsi
                - tanggal
                - lokasi
                - foto
              properties:
                nama:
                  type: string
                  maxLength: 50
                deskripsi:
                  type: string
                tanggal:
                  type: string
                  format: date
                lokasi:
                  type: string
                  maxLength: 255
                foto:
                  type: string
                  format: binary
              type: object
      responses:
        201:
          description: 'Berhasil dibuat'
          content:
            application/json:
              schema:
                properties:
                  data:
                    $ref: '#/components/schemas/Documentation'
                type: object
        422:
          description: 'Error validasi'
        401:
          description: Unauthorized
      security:
        -
          bearerAuth: []
  '/documentations/{id}':
    get:
      tags:
        - Documentations
      summary: 'Menampilkan detail dokumentasi'
      operationId: 8f92160da5af359f5035faf9e348fd11
      parameters:
        -
          name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        200:
          description: Berhasil
          content:
            application/json:
              schema:
                properties:
                  data:
                    $ref: '#/components/schemas/Documentation'
                type: object
        404:
          description: 'Not found'
    put:
      tags:
        - Documentations
      summary: 'Memperbarui dokumentasi'
      description: "Memerlukan autentikasi dan role 'admin'. Dapat mengganti foto."
      operationId: 387dd643b5fadd1e031c0a3710038b90
      parameters:
        -
          name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          multipart/form-data:
            schema:
              properties:
                nama:
                  type: string
                  maxLength: 50
                deskripsi:
                  type: string
                tanggal:
                  type: string
                  format: date
                lokasi:
                  type: string
                  maxLength: 255
                foto:
                  type: string
                  format: binary
              type: object
      responses:
        200:
          description: 'Berhasil diperbarui'
          content:
            application/json:
              schema:
                properties:
                  data:
                    $ref: '#/components/schemas/Documentation'
                type: object
        404:
          description: 'Not found'
        422:
          description: 'Error validasi'
        401:
          description: Unauthorized
      security:
        -
          bearerAuth: []
    delete:
      tags:
        - Documentations
      summary: 'Menghapus dokumentasi'
      description: "Memerlukan autentikasi dan role 'admin'"
      operationId: 9f58ce8dc08de6bcfc6ee100121cae99
      parameters:
        -
          name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        200:
          description: 'Berhasil dihapus'
        404:
          description: 'Not found'
        401:
          description: Unauthorized
      security:
        -
          bearerAuth: []
...
components:
  schemas:
...
    Documentation:
      required:
        - nama
        - deskripsi
        - tanggal
        - lokasi
      properties:
        id:
          type: integer
          example: 1
        nama:
          type: string
          maxLength: 50
          example: 'Kegiatan Bakti Sosial'
        deskripsi:
          type: string
          example: 'Kegiatan sosial di desa Sukamaju.'
        tanggal:
          type: string
          format: date
          example: '2024-10-20'
        lokasi:
          type: string
          maxLength: 255
          example: 'Aula Desa Sukamaju'
        foto:
          type: string
          example: documentations/nama_file.jpg
        foto_url:
          type: string
          example: 'http://localhost:8000/storage/documentations/nama_file.jpg'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      type: object
...
  securitySchemes:
    bearerAuth:
      type: http
      name: Authorization
      in: header
      bearerFormat: JWT
      scheme: bearer
...